// agentark-web/src/shared/api/generated/control/models/cursorPageSecretBinding.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AgentArk Public Control API
 * OpenAPI spec version: 1.0.0
 */
import type { CursorPageSecretBindingItemsItem } from "./cursorPageSecretBindingItemsItem";

export interface CursorPageSecretBinding {
  items: CursorPageSecretBindingItemsItem[];
  /** @nullable */
  nextCursor: string | null;
  hasMore: boolean;
}