// agentark-web/src/shared/api/generated/control/models/cursorPageSecretMetadata.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * AgentArk Public Control API
 * OpenAPI spec version: 1.0.0
 */
import type { CursorPageSecretMetadataItemsItem } from "./cursorPageSecretMetadataItemsItem";

export interface CursorPageSecretMetadata {
  items: CursorPageSecretMetadataItemsItem[];
  /** @nullable */
  nextCursor: string | null;
  hasMore: boolean;
}